PANDO-39: Fixes for BaseSerializerTest for variable size serializers
For serializers for which the number of bytes to deserialize was based on the contents of the buffer, the current test implementation would cause issues because it would use an empty buffer, which could change the number of byte to deserialize, thereby causing erroneous results
This now actually uses the "real" bytes to deserialize, but manipulates the length of the buffer to get the results we want.

This was instigated by implementing tests for NullableSerializer, which changes the number of bytes it reads/writes based on the value.

work 0h 12m