PANDO-39: Add overridable comparison of deserialized values
The AssertDeserializedValuesEquivalent method allows subclasses to override the logic for asserting that the deserialized value matches the expected value. This can be useful for cases where the default Equals implementation for a value is not sufficient to completely assert that values are equivalent (e.g. DateTime and its Kind property)