Introduce IPandoSerializer
IPandoSerializer unifies the old INodeSerializer and IPrimitiveSerializer. Unlike the old system, implementations of IPandoSerializer don't have to know anything about whether its children are nodes or primitives, as the child serializer is responsible for writing whatever data is appropriate into the buffer.

This paves the way for GenericNodeSerializer, which can serialize nodes without needing an explicit implementation, removing the need for the Pando source generator completely and paving the way for the implementation of Merge